updated parser and added supplements
