--- src/main/java/com/vp/amazonreviewsapp/util/Parser.java
@@ -1,5 +1,0 @@
-package com.vp.amazonreviewsapp.util;
-
-public interface Parser {
-    void parseData(String path);
-}
