+++ src/main/java/com/vp/amazonreviewsapp/util/parser/Parser.java
@@ -1,0 +1,7 @@
+package com.vp.amazonreviewsapp.util.parser;
+
+import java.util.List;
+
+public interface Parser {
+    List<List<?>> parseData(String path);
+}
